+++ agenta-web/src/lib/Types.ts
@@ -165,3 +165,8 @@ export interface ISession {
 }
 export type GenericObject = Record<string, any>
 export type KeyValuePair = Record<string, string>
+
+export interface Environment {
+    name: string
+    deployed_app_variant: string
+}
